On branch master
 Changes to be committed:

added a text field and corresponding code in server.R so that a specific
password resets everything.  The password is 'el'

added code so that when the number of players is reset from a higher
number to a lower number the data for the higher numbered players is
reset, e.g. when the number of players is reset from 2 to 1, the data
previously entered for player 2 is deleted

	modified:   server.R
	modified:   ui.R